lint: use same type for comparison

+++ src/builtins/unset.c
@@ -6,16 +6,15 @@
 /*   By: Oery <coincoin@baozi>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/10 23:56:01 by Oery              #+#    #+#             */
-/*   Updated: 2026/03/10 23:58:56 by Oery             ###   ########.fr       */
+/*   Updated: 2026/03/11 18:30:52 by Oery             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-#include <stddef.h>
 
 int	mini_unset(int argc, char **argv, t_ctx *ctx)
 {
-	size_t	i;
+	int	i;
 
 	i = 1;
 	while (i < argc)
